Eliminate 'comparing signed/unsigned values' lcc warning in add_chunk_as (a cherry-pick of commit c56149c7e from 'master')
* src/atomic_ops_malloc.c (add_chunk_as): Cast CHUNK_SIZE to size_t
when compared to value of sz local variable.

+++ src/atomic_ops_malloc.c
@@ -268,7 +268,7 @@ static void add_chunk_as(void * chunk, unsigned log_sz)
   size_t ofs, limit;
   size_t sz = (size_t)1 << log_sz;
 
-  assert(CHUNK_SIZE >= sz);
+  assert((size_t)CHUNK_SIZE >= sz);
   assert(sz % sizeof(AO_t) == 0);
   limit = (size_t)CHUNK_SIZE - sz;
   for (ofs = ALIGNMENT - sizeof(AO_t); ofs <= limit; ofs += sz) {
